Updated Phoenix, fixed all that crushed following the changes, ehich is that now getSelectedSensorPosition returns doud not int

+++ src/sensors/counter/CtreEncoder.java
@@ -21,7 +21,7 @@ public class CtreEncoder implements Counter {
 
 
   @Override
-  public int getCount() {
+  public double getCount() {
     return baseTalon.getSelectedSensorPosition(pidSlot);
   }
